pyproject.toml
[project]
name = "someproject"
version = "0.1.0"

dependencies = [
  "speechtotext-python",
]

[tool.setuptools.packages.find]
include = [
  "someproject",
]
